Trying to fix animations
idk what is happening the gun ik's for the hands are real screwy, no matter what ive tried it doesnt seem to want to properly put the sword in place properly for the ready stance. even though the idle stance works and the running stance. i dont get it, though ill keep trying to get the poses right.