New Battle Lock System
In the second level the player has to fight an entire village worth of enemies. I decided to go more devil may cry style and make a mandatory battle event.
Once the player is in the trigger walls appear and stop the player from getting out, the only way to continue is by beating all the enemies spawned by the battle lock trigger
this also works with saves